% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_play_stats.R
\name{add_winner_data}
\alias{add_winner_data}
\title{Add Winner Data}
\usage{
add_winner_data(play_df)
}
\arguments{
\item{play_df}{a data frame of gameplay statistics}
}
\value{
the same play_df with an appended winners column
}
\description{
Takes a play df and adds a column of winner data.
}
\details{
Takes a dataframe of gameplay stats as input. Finds all the winners
for that game's id, and makes a list of them. Appends that list back
to the original data frame.
}
