% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_index}
\alias{find_index}
\title{Find Index for Gaming Statistics}
\usage{
find_index(num_vect)
}
\arguments{
\item{num_vect}{A numeric vector}
}
\description{
A helper function to count for N x N gaming indeces.
}
\details{
Takes in a numeric vector for which we're looking to find a N x N index.
Returns the highest N where N >= N items deep in a descending sorted list.
}
