% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_play_stats.R
\name{generate_game_stats}
\alias{generate_game_stats}
\title{Generate Game Statistics}
\usage{
generate_game_stats(play_df)
}
\arguments{
\item{play_df}{a data frame of valid plays}
}
\description{
Get statistics for all the unique games played.
}
\details{
Input a play df. Return a data frame of statistics characterizing the play
patterns of each individual game.
}
