% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_play_stats.R
\name{generate_player_stats}
\alias{generate_player_stats}
\title{Generate Player Statistics}
\usage{
generate_player_stats(play_df)
}
\arguments{
\item{play_df}{a data frame of valid plays}
}
\description{
Get statistics for all the individual players.
}
\details{
Input a play df. Return a data frame of statistics characterizing the play
patterns of each individual player.
}
