% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_play_stats.R
\name{get_gaming_index}
\alias{get_gaming_index}
\title{Get Z x Z Index}
\usage{
get_gaming_index(play_df, play_index)
}
\arguments{
\item{play_df}{a data frame of valid plays}

\item{play_index}{an index of any of the index values: "H", "HH", "K", or "KH", not case sensitive}
}
\description{
Get a gaming index, such as the H-index.
}
\details{
Input a play df and the index you desire.
H-index: N where you have played N distinct games at least N times each.
HH-index: N where you have played N distinct games for at least N hours each.
K-index: N where you have played with N distinct people for at least N games each.
KH-index: N where you have played with N distinct people for at least N hours each.
Returns the value of the given index.
}
