% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_bg_plays.R
\name{play_xml_to_df}
\alias{play_xml_to_df}
\title{Turn a Board Game Geek file into an R Data Frame}
\usage{
play_xml_to_df(xml_play, player_data = "players", game_data = "item")
}
\arguments{
\item{xml_play}{An individual BGG play, usually one element of an XNL list.}
}
\value{
A data frame of gameplay statistics.
}
\description{
Takes a Board Game Geek xml list parameter. It outputs a data frame where all
plays are sorted by play_id (listed in the data fram as `id``) and date.
}
