% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_bg_plays.R
\name{playXML_to_df}
\alias{playXML_to_df}
\title{Turn a Board Game Geek file into an R Data Frame}
\usage{
playXML_to_df(xml_play, player_data = "players", game_data = "item")
}
\arguments{
\item{xml_play}{The BGG play.xml file that is loaded into R.}

\item{player_data}{The name of the xml sub-node where player data is kept. "players" by default.}

\item{game_data}{The name of the xml sub-node where game data is kept. "item" by default.}
}
\value{
A data frame of gameplay statistics.
}
\description{
This function takes a Board Game Geek xml list parameter It outputs a
data frame where all plays are sorted by play_id (listed in the data fram as `id``) and date.
}
