% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_bg_plays.R
\name{read_bgg_stats_xml}
\alias{read_bgg_stats_xml}
\title{Read a BGG play.xml file into R}
\usage{
read_bgg_stats_xml(
  user,
  base_url = "https://boardgamegeek.com/xmlapi2/plays?username="
)
}
\arguments{
\item{user}{The username of a Board Game Geek account.}

\item{base_url}{The url of Board Game Geek's xml api.}
}
\value{
A list of gameplay statistics in XML style.
}
\description{
This function takes an xml website and username as inputs. It looks at
all of the logged games from BGG as listed by the xml website, and loads those
into R as a list. It outputs a list that has the same structure as the xml file.
}
