% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_bg_plays.R
\name{clean_playXML}
\alias{clean_playXML}
\title{Consider the XML Only from Valid Games}
\usage{
clean_playXML(xml_play, player_data = "players", game_data = "item")
}
\arguments{
\item{xml_play}{A BGG play.xml file that is loaded into R.}
}
\value{
Two  possible outcomes. 1. A BGG play.xml file that is stripped of invalid plays. 2. An error that says "no valid plays logged"
}
\description{
This function takes in and returns a Board Game Geek xml list parameter.
The output file will only have valid plays.
}
